eayunrgw: haproxy stick-table size should equal to 2 when stick on dst.
see the comments on page:
http://blog.haproxy.com/2014/01/17/emulating-activepassing-application-clustering-with-haproxy/

note:
the comments in above page said stick-table size should set to 1, but in
fact when setting to 1, if the backend server of the existed entry is
down, new connections will be blocked as the existed entry would not
refresh. Setting to 2 will avoid this problem.

Signed-off-by: Zhao Chao <zhaochao1984@gmail.com>